--- carma_nav2_behavior_tree/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>carma_nav2_behavior_tree</name>
-  <version>0.1.0</version>
-  <description>Behavior tree nodes to supplement those in nav2_behavior_tree</description>
-  <maintainer email="carma@dot.gov">CARMA Support</maintainer>
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>nav2_common</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>nav_msgs</depend>
-  <depend>nav2_behavior_tree</depend>
-  <depend>nav2_msgs</depend>
-
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_cmake_gtest</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
