--- carma_nav2_emergency_stop/README.md
@@ -1,3 +1,0 @@
-# `carma_nav2_emergency_stop` package
-
-This package adds emergency stop functionality to shutdown the Navigation2 stack and send a halt command when the center button on the Logitech Gamepad F710 is pushed.
