--- carma_nav2_emergency_stop/package.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>carma_nav2_emergency_stop</name>
-  <version>0.1.0</version>
-  <description>An emergency stop for navigation2</description>
-  <maintainer email="carma@dot.gov">CARMA Support</maintainer>
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>nav2_msgs</depend>
-  <depend>nav2_util</depend>
-  <depend>rclcpp</depend>
-  <depend>rclcpp_lifecycle</depend>
-  <depend>sensor_msgs</depend>
-  <depend>nav2_lifecycle_manager</depend>
-  <depend>ackermann_msgs</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
