--- carma_nav2_port_drayage_demo/package.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>carma_nav2_port_drayage_demo</name>
-  <version>0.1.0</version>
-  <description>A port drayage demo navigation2 application</description>
-  <maintainer email="carma@dot.gov">CARMA Support</maintainer>
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <build_depend>ament_auto_cmake</build_depend>
-
-  <depend>carma_v2x_msgs</depend>
-  <depend>nav2_msgs</depend>
-  <depend>nav2_util</depend>
-  <depend>rclcpp</depend>
-  <depend>rclcpp_lifecycle</depend>
-
-  <build_depend>nlohmann_json_dev</build_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_cmake_gtest</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
